User impact: read edit count from primary db in save complete hook
Use flag READ_LATEST in the ImpactHooks::userIsInImpactDataCohort check
and in the call to refreshUserImpactDataInDeferredUpdate so we ensure
the data is refreshed after the user makes an edit and the
data refreshed includes the edit just made.

This prevents new created accounts from finding a broken/empty impact
module after they make their first edit.

Bug: T324930
Depends-on: Id7223ee4953a3e066d6028a41072c7369ea1ce88
Change-Id: I6bd0daf5e77ca5f1ff82983f830bb01f3acab9cd