WikiPageConfigLoader: Stop treating "page does not exist" as failure state
The intended behavior of "config page does not exist"
and "config page is empty" are equal. In both cases,
GrowthExperiments are configured based on PHP globals
(or consider the mentor list empty, depending on which
user of the WikiPageConfigLoader we're considering).

For that reason, let's stop treating
the "page does not exist" case as a failure
state, and pretend the page is empty instead.

Bug: T325236
Change-Id: Iadb4e24a786605c54374f76694e845d00c3785b3