Vue components: react to binding updates of v-click-outside directive
The clickOutside directive was ommiting updates in the binding.value
which are relevant when the value bound is a reactive property.

CPopover component was passing a close event handler to the clickOutside
directive regardeless of its state being open or close making
unnecessary calls for all closed instances. This allows to consistently
listen to @toggle events. Add also @open and @close events
for convenience.

Bug: T325041
Change-Id: I854631a1fe7589e1fb35bb758bf60d51f4d85fbf