Vue components: refactor CText to use standard template
Using the render function inside a single-file component with an initial
template impedes @vue/test-utils to wrap the result of the function
when calling mount(Component) and it returns the content of the template.
Removing the template tag is not possible either as it yields linting
errors  and ResourceLoader fails to load the file with No <template> tag
found.

Since render was only used for the dynamic creation of the tag (prop as), this  can be done with template syntax <component is=p></component> easing
the calls to mount in tests without breaking RL loading.

The change alters the test results since CText renders its slot content
as opposite of just <div></div>

Bug: T324549
Change-Id: I500b23cd1f27f89abfe8b1157e7f04d987b0e205