Add test for UnderlinkedFunctionScoreBuilder
Add test for Ia11e92dc75c96e9ca3687faef043e53e9b1483cb.

Not straightforward because the interesting part of
UnderlinkedFunctionScoreBuilder is a JS snippet (and a proper
integration test would require setting up ElasticSearch) so we
just copy over the JS code to a test file.
Put it in api-testing to avoid the hassle of setting up JS unit
tests just for this.

Bug: T301096
Change-Id: I37e3f4ea2c52ecc6009889c8e6d20d1bed51066c