User impact: add skeleton markup while JavaScript loads
Place styles of components used in the server first render in
less files. Create some skeleton CSS utility classes to display
minimal content skeleton while JavaScript is loading.

The code could still be adjusted to display better skeletons for
the unactivated state and the mobile summary. In the full overlay
rendering an skeleton is not done since the view is not presented
until JS loads.

Bug: T321675
Change-Id: Ic98d374c25685cab26aa1821576cc8183e82f7c1