ChangeMentor: If mentee has mentorship module disabled, enable it
At eswiki and enwiki, mentorship is not yet scaled
to 100% of mentors. Ocasionally, mentors want to claim
a mentee, who does not actually have access to mentorship.

In that case, enable the mentorship-module for the mentee.

This ignores mentees who intentionally opted-out of mentorship
(in other words, such mentees will not be given module back
automatically; they'd have to opt-in manually via their Homepage).

Bug: T327206
Change-Id: I666d060a2120b65b9cdc9329eb3daf3e8ed0e7be