Suggested edits: use existing FiltersButtonGroupWidget update method
FiltersButtonGroupWidget provides an updateButtonLabelAndIcon method to
handle the header state. In the prior patch we missed the method and
added an ad-hoc update of the icon when the dialog closes. While that
works fine the updateButtonLabelAndIcon is used in other updating
operations so we would benefit to have it in a single place.

Bug: T318597
Follow-up: I8aaed8d944ca80534bfb59d4f3d03698001045a7
Change-Id: I9487d7135cf431f11fa618f7efb3c2c701cfb5b8