Post edit panel: show the task description skeleton while requesting the PCS
The SmallTaskCard widget reads the task.description property and shows
a skeleton for it when it is undefined. Initializing it to null was preventing
it to render in the post edit dialog while we request the PCS.

Note that the description does show in the mobile homepage because it is
rendered on the server in SuggestedEdits::getTaskCard.

Bug: T294835
Change-Id: Iba37c3dda93c76309947deaf3bd1e77741256e58