[docs] Improve several VisualEditorHooks-related comments
Since 6ad99ddd7f, we have a hard dependency on VisualEditor.
Because that happened, remove the "this is because VE is an optional
dependency" comment from VisualEditorHooks (we could likely put it back
to HomepageHooks if we wanted to).

Update some JS comments that were pointing to the hook handler
in its older location.

Change-Id: If406b890635a6e3aaf5788d51614de23e1d410ee