Soften dependency on EventBus
If EventBus is not loaded, AddImageSubmissionHandler
should not send any events to it. This unbreaks
Growth features on non-EventBus wikis.

An alternative solution would be formally
declaring a hard dependency on EventBus.
Historically, the team norm is to avoid introducing
hard dependencies whenever possible, and in this case
EventBus doesn't seem to be a strictly needed
dependency.

Another alternative solution would be to
introduce a config variable similar to
GELinkRecommendationsUseEventGate, but that doesn't seem
to bring in additional benefits.

Follow-up to I0d2f5f977254b2d4329f54b6682726a87066b98c.

Bug: T329672
Change-Id: Ic619318a1d74aaf5b34eaab1346231642bc03eba