Leveling up: Show invite dialog after non-suggested edits
Add a ext.growthExperiments.LevelingUp.InviteToSuggestedEdits module
for showing a dialog after select non-suggested edits that invites
the user to do suggested edits. Load the dialog on postEdit and
VE save when conditions match.

Also improve the documentation of some SuggestedEditSession methods
(which in the end didn't need to be changed).

Handling mobile edits is left for a follow-up.

Testing notes:
* Set something like
      $wgGELevelingUpManagerInvitationThresholds = range( 1, 100 );
  so all edits trigger an invitation.
* You can also trigger the dialog without editing by loading an
  article with the venotify=saved query parameter.

There are in theory six meaningfully different code paths for a
post-edit notice:
* wikitext editor save with post-redirect-get
* VE save that refreshes article content without page reload
* VE save that triggers a page reload via JS (after switching
  from the wikitext editor, or when creating a new page or
  opening VE from the page history)
* mobile wikitext editor save that uses the API and then triggers
  a page reload via JS
* mobile VE save that does trigger a reload (when creating a
  new page)
* mobile VE save that changes the URL fragment but does not reload

This patch does not cover the last scenario.

Bug: T322387
Change-Id: Ib8e953f9200773ce2a2af7fb9831648229e0ca6b