Leveling up: Suggested edits invitation server side
Add $wgGELevelingUpManagerInvitationThresholds and
LevelingUpManager::shouldInviteUserAfterNormalEdit
for checking whether suggested edits should be
recommended to a new editor.

Set a flag on the VisualEditor save API response in
non-suggested edits when the user is eligible to be
invited to try suggested edits.

Also get rid of SuggestedEdits::SUGGESTED_EDIT_TAG
which wasn't used anywhere.

Handling of non-visual edits is left for a follow-up
as a similarly clean solution is not available: either
we'd need to use the GET request and deal with replag
and the fact that there is no clean way to identify
in a GET request that we are just after an edit,
or use the POST and pass preserve the information
across the post-redirect-get cycle and use a sub-par
hook (as PageSaveComplete is deferred and can't
affect the response), or use a standalone API request
unrelated to saving the article, which would then be
an added complication for all article edits unrelated
to GrowthExperiments features.

Testing notes:
Set something like
    $wgGELevelingUpManagerInvitationThresholds
      = range(1, 100, 2);
which makes every odd-numbered edit eligible for
invitation. After that, the API response should have a
visualeditoredit.growthexperiments.levelingup.invited
field during saving the edit if it is eligible.

Bug: T322387
Change-Id: I157590e31acfbc3b43134f0c37f402e6718a4bc4