API: Add "growthnextsuggestedtasktype" query module
Add an API query module for wrapping the functionality of the
LevelingUpManager service. This can be used by client-side code to ask
for a suggested next task type for a user. To avoid confusion with
database replication issues, it should be called before
the user makes an edit, when the user begins to work on a particular
task (e.g. the article page has opened).

Bug: T322386
Change-Id: I3a096afb837cc359c103966eb102f6ac3a614130