Build backend for PersonalizedPraise
This patch adds logic to determine whether or
not a mentee is praiseworthy.

For simplicity, this patch reuses UserImpact
value object, instead of adding its own (which
would be a strict subset of UserImpact).

The patch consists of two things:

* Getting the list of praiseworthy mentees
* Ensuring the list can be fetched quickly via the UI

==== Getting the list of praiseworthy mentees
This happens at the application level. PraiseworthyMenteeSuggester
iterates through all active mentees mentored by a given mentor.
For each of them, it decides whether they are praiseworthy
or not based on data fetched from UserImpact. I decided to use existing
data fetching logic, because I did not want to be reimplementing more or less
the same thing again in the same extension.

==== Ensuring the list can be quickly fetched
On each edit of the mentor's mentees, a check
for praiseworthiness is executed. If it passes,
the mentor's list of praiseworthy mentees is updated
in a cached storage.

I decided to use the main stash to cache
the list of praiseworthy mentees. This decision
was made, because the main stash is semi-permanent
(as regeneration cannot be triggered by the mentor).

Bug: T322444
Depends-On: I06c625abe351b2f189f5dbf6f331d34bb56004b2
Depends-On: I978fefa239335027747cc26c2c7f1f364c151604
Change-Id: Ib7b495c9b04b06f495d53d9b3b7a1656bc6fae29