Do not enable mentorship when mentor is changed in bulk
This also skips notification for bulk mentor reassignment.

As of T327206, newcomers who are claimed by a mentor are guaranteed
to receive access to mentorship module, even if they previously
had MENTORSHIP_DISABLED. This is to ensure mentors can claim users
freely, without wondering if they'd see the change in some way.

Unfortunately, the T327206 change affected bulk reassignment
which happens after a mentor quits. Avoid this by introducing
a concept of bulk mentor changes. As of now, only treat
mentor changes triggered by a mentor resignation as bulk. In the
future, we might have to consider exposing this concept from
ApiSetMentor as well.

All bulk changes of mentors proceed as normal ones, with two
differences:

	1) ChangeMentor does not enable mentorship module for
	   bulk changes.
	2) If newcomer doesn't have mentorship access after the change
	   (because of 1), notification is skipped.

Bug: T301981
Change-Id: I629d7380c1abbec616d0a837e40722e7b15f470f