Vue components: test CPopper
Adding tests to CPopper as a way to document the valid props input.
The icon label prop is only "required" when an icon is being provided.
Codex already throws a warning for this and the unit test suite exits
with error if a warning is yield (see jest.setupAfterEnv.js), hence not
adding a custom validator.

Bug: T324549
Bug: T326233
Change-Id: I9d4bb1c15a9152d7945ecc3aa7d480010998f34b