User impact: add missing icon label for close button in ScoreCard component
Also make icon mocks truthy in tests so they are rendered and add
missing test case to render the popover content in the ScoreCard test.
Mocking the icons to empty strings may cause some components to not
render them if there's a v-if check. Only taking action in the
cdxCloseIcon in this patch.

Bug: T326233
Change-Id: I88cbbb7228c373ad0873298949ce25318e18a3dc