Leveling up: InviteToSuggestedEdits instrumentation
Instrument impression, close, button click, collapse, expand
events of the non-suggested post-edit panel, and homepage
visits originating from the panel.

Also update HelpPanelLogger to treat post-edit events as editing
for privacy purposes (ie. log what article we are on). This
affects instrumentation for existing dialogs as well.

Also update HelpPanelLogger to set 'not-known' as editor when it
can't guess the current editor. This also affects existing
instrumentation (previously we just guessed 'visualeditor').
We'll need it because the module is only loaded after the save
happens, and previous logging implementation relied on the logger
being present in the editing process and storing the editor type,
but for non-suggested edits we try to be less invasive and only
load when needed.

Bug: T330727
Change-Id: I37dbd38eb5f14e0ef4c47a6435a184b7cc40932f
Depends-On: I59bdf17696d017c010799c79bdfee3f0f089a213
Depends-On: I1618795611596a85bcbb2a9f1d0262af5611a81d
Depends-On: I70bc894b9eb5b900a6b212203a5abc4bc2225879