TourHooks: Set new impact discovery to seen for all created accounts
The new impact module tour should be declared as seen for all newly
created accounts, whether or not they are auto-created or local, and
regardless of whether they have homepage enabled.

In TourHooks::onUserGetDefaultOptions we make sure that existing user
accounts can see the tour, by declaring the default "seen" flag for this
tour to false.

Bug: T330692
Change-Id: I1354325e47794637f2d53b18a111384c5018ef2e