Leveling up: always set wgGELevelingUpEnabledForUser
wgGELevelingUpEnabledForUser is needed in the post-edit dialog.
We either need to show it on every pageview (as we can't tell on
the server side when the user is in a suggested edit session) or
need to persist it as a property of SuggestedEditSession. I went
with the simpler solution here.

Bug: T332227
Change-Id: I5f2c369c7476ce40e28b7eb15d8eac4fd9c799db
(cherry picked from commit c60dc9006f5a7e0c01efe2a5b27edcb22631efe4)