LevelingUpManager: Handle links/link-recommendation collision
We allow links and link-recommendation to co-exist as task types because
we used to have an A/B test, where some users would see one and others
would see the other.

On wikis where link-recommendation is enabled, that means that
ConfigurationLoader::getTaskTypes() will return 'links' even though that
is not a valid task type for users to interact with. The same situation
applies in reverse for wikis where link-recommendation is not enabled.

A better way to handle this might be to enable/disable the task types
in PageConfigurationLoader (T332348), but that is a farther reaching
change, so go with a hack for now.

Bug: T332309
Change-Id: I94f6eff3674b9b27570df8d971d18573db225119