Rename CdxButton `type` prop to `weight`
In Codex v0.7.0, the CdxButton's `type` prop changed. The `type`
prop was renamed `weight` and will continue to accept the same
values ('normal', 'quiet', and 'primary'). A new `type` prop was
added that accepts the old types, plus native button types
('button', 'reset', and 'submit'). In a future release of Codex,
the `type` prop will be removed altogether and the native type
can be bound to the CdxButton via the `type` attribute.

So, while the old values for `type` still work, we want to
update the prop name in anticipation of the `type` prop's
removal. For any new code moving forward, the `weight` prop
should be used.

This patch:
- Renames all uses of the `type` prop to `weight`
- Removes one instance of `weight="primary"` that was not paired
  with the progressive or destructive action (primary weight has
  no effect without one of those actions)
- Changes `null` to `undefined` in a couple of ternaries in the
  template to ensure the proper button classes are output
- Updates Codex dependency to 0.7.0

Bug: T312987
Change-Id: I480386b0053d2841be0992074437fec87871d3c6