PostEdit: Increment the edit-count-for-task-type count
We currently fetch the "next task type" when the article loads, and when
the post-edit dialog appears. That code also updates the SuggestedEdit
Session's "editCountByTaskType" property. That means that when the
try new task panel appears, the editCountByTaskType is not updated and
is thus off-by-one.

We could avoid this +1 hack by blocking the UX on the completion of
LevelingUpManager's ::suggestNewTaskTypeForUser method:

1. Our VisualEditorHook PostSaveHook could return back the next
   suggested task type and the edit count by task type, so no API
   request is necessary. But we'd need to adjust VisualEditor to include
   the data object when it fires `mw.postEdit` hook, and we'd still need
   to handle the source editor separately.
2. We could not show the post-edit panel or the try-new-task panel until
   the API request to growthnextsuggestedtasktype completes, but this is
   problematic because depending on how many topics the user has
   selected, the query could take some time. Plus it is another
   roundtrip HTTP request for the client to make.

We could explore either of those options in a follow-up.

Bug: T332319
Change-Id: Id8e434eaa269005f2479340bedf526a520b99713