GrowthMentors.json: Add a write-only username field
This adds a new optional field to MediaWiki:GrowthMentors.json,
called `username`. The newly added field will be used by administrators
making manual edits to the mentor list, so they can easily identify
an entry for a particular user they're interested in.

Apart from manual editors, this newly added field is not,
and SHOULD NOT, be used for anything else. The only identifier
used by structured mentor list must be the user ID, to avoid
issues with renames etc.

This patch does not handle renamed users. It may or may not
be possible to hook into a rename and edit the list using
a system account. Renames might be handled in the future
within T332410.

This patch will cause issues if a train is rolled back (mentorship
would be unusable, as there'd be an unexpectable key in the mentor list)
Removing the field manually is the treatment plan for that.

Bug: T331444
Change-Id: I87b47675a097d340ca5716fb19089ee4e8dbde73
(cherry picked from commit f5630fe691584f4c0db49409f088c72dbb76193e)