Personalized praise: Make it possible to deploy without notifications
Initially, we want to deploy the Personalized praise
module to Growth pilot wikis with no support for
notifications. Make that possible by config variables.

GEPersonalizedPraiseEnabled's usage in Community configuration
is replaced by GEPersonalizedPraiseBackendEnabled. I'd prefer tying
the Community configuration appearance with whether the frontend is enabled,
but I don't want to introduce yet another global just for that (plus,
the actual frontend apperance wouldn't be governed by that variable anyway).

Ideally, Mentor dashboard would expose which modules it has enabled,
but that's not possible as of now.

Bug: T334630
Depends-On: I32c86e922b61dfdb16af2337aaa198f33467c3a1
Change-Id: I294a604449b7aec72960f3efb36b9219bab34926