CScoreCards: Move logging responsibility upper in the stack
Previously, CScoreCards contained a call of
	$log( 'impact', 'open-thanks-info-tooltip' )

This worked fine in Homepage, as the $log
function was provided by the NewImpact
module. However, in Mentor dashboard,
there is no $log function injected and as such,
console errors were logged.

Resolve this by moving the logging responsibility
higher up the stack, under NewImpact.vue. In that component,
the existence of $log can be safely presumed.

Bug: T335614
Change-Id: I28ea7b75f0c4bdfab8e99c64a787cc353b8c9253