Section images: update growthinvalidateimagerecommendation
Allow passing a task type to
AddImageSubmissionHandler::invalidateRecommendation so it can
invalidate both top-level and section-level recommendations.
This requires adding a task type to the validate()/handle()
interfaces.

Have the action=growthinvalidateimagerecommendation API support
section-level image recommendations by taking the task type as a
parameter.

Also includes a small code change to the growthimagesuggestiondata
API to keep the code of the two APIs similar.

Bug: T329278
Change-Id: I32c9e1cb2bd35ff9e2cf6cca140d7759d3791c8a