Allow some interactions with disabled task types
When a task type is disabled (via the $wg...Enabled flag  or the
'disabled' community configuration property), keep hiding it from
the user but allow the growthimagesuggestiondata API to access it.
This is useful for other clients (such as the Android app) and for
ActionApiImageRecommendationApiHandler on the beta cluster.

To facilitate this, provide ConfigurationLoader::getDisabledTaskTypes().

Bug: T329278
Bug: T336358
Change-Id: I00e99012eefab208339c75b5c28beadc2782e3c0