addsectionimage: Introduce quality gate infrastructure
Similar to I9132342662c289945b25011091489e21a2bf8b7c, but stops short of
implementing the submission handling logic; we will implement that in a
follow-up.

For now, this patch gets various pieces in place so that we don't have
JS errors when visiting a section-level image article from Special:Homepage.

Bug: T336550
Change-Id: Ieab523b0f10836c6197ab743f0e1ae5d005724b3