New impact: show call to action in mobile-details mode
Add support for mobile-details render mode in the client module. Prior
to this change the render mode mobile-details was not handled
by the Vue application and treated as if it was a mobile-overlay mode.
Rename mode names used in the client module to match server ones.

Show the suggested edits call to action in NoEditsDisplay component
in both mobile-overlay and mobile-details modes.

Also:
 - Move some LayoutWrapper styles to a less file so they can
be used while the skeleton displays.
- Remove no longer needed mode conversion in the instrumentation
logger.

Bug: T337314
Change-Id: I01c4f2ee9c330f6d75dc4a6f3d60c4565c9d2512