ReassignMentees: Drop mentor/mentee relationship when no mentor is available
When no mentor is available for reassignment (happens when
the last mentor quits), mentorship should be fully disabled,
instead of displaying a mentorship module with an ex-mentor
in it to the newcomers.

To do that, the mentor/mentee relationship is dropped from
the database if the last mentor resigns. Homepage will then
fail to assign a new mentor, and hide the mentorship module.

In addition to this change, we might want to consider
automatically setting GEMentorshipEnabled to false
via Community configuration.

Bug: T327399
Change-Id: I51968773ae5cea685257bc8d17ba29cb26acb55f