Section images: Accept all API response 'kind' fields
Since 'kind' values in the production image-suggestions API
responses are unpredictable at this point, accept any 'kind' value
and use the presence of a section title instead to determine
whether it is a section-level recommendation.

Still try to use the 'kind' field to select an appropriate i18n
message explaining where the recommendation comes from, but select
a generic one if the field value is not recognized.

Also make sure the ranking of recommended images uses a stable sort
as that was causing issues with a test.

Bug: T329278
Bug: T337702
Change-Id: Iafadfee0fa3b55bba2dfa479f5bfa26d3a4c43af