Section images: Remove user variant
We won't do an A/B test so remove the section images user variant
and related code.

Reverts part of I1affbc50146815. Keeps the feature flag, and also
the user variant related refactoring (which isn't necessary anymore
but IMO makes the code nicer and means it's less work if we do
decide to make an A/B test for some other future task type).

Bug: T329276
Change-Id: Ie5afa618e34b1d58939883bac193c89b59cb808b