Section images: Implement image placeholder
Implement the "placeholder" for section images where we show
an icon to indicate where the image will be inserted, before
user approval of the image (when insertion actually happens).

There are two ways to achieve this: make the placeholder a
subclass of MWImageNode (either by having RecommendedImageNode
act as both placeholder and actual image, depending on some
flag that's managed as a node attribute; or by having separate
image and placeholder node types, both MWImageNode subclasses)
or make it a custom node that's not derived from VE types.

This patch does the latter, on the grounds that the placeholder
will have a fairly different DOM from the actual image, only
needs very simple data model + contenteditable behavior
(although most behavior is blocked by readonly mode anyway so
that doesn't matter much), and the effort needed to simulate
the image node's layout and positioning isn't large.
This means having to reimplement skin/device specific image
formatting, but not having to deal with unintended consequences
of having a DOM node that uses image thumbnail classes but
doesn't exactly have the same structure and expected design as
actual images.

Other changes:
* Fix specificity of image node max-width rule, which was
  getting overruled on mobile.

Fixing the scrolling / transitions is left for another patch.

Bug: T335209
Change-Id: Ie400067723eb7fad94dd26a058b1361aaa1210ee