Section images: Fix scrolling behavior
Adjust scrolling behavior so that the relevant section is
brought up when the toolbar dialog is displayed.

Creates a unified way of triggering scrolling behavior,
StructuredTaskToolbarDialog.scrollToToolbarDialogTarget. This
will be used by a follow-up patch to scroll back to the section
when the toolbar dialog is minimized and restored.

Misc:
* Improve some unrelated docstrings.

Bug: T335209
Change-Id: Ib682fed13eb939daa454df64ca8b56240baf418c