Add image: Don't scroll on desktop in caption step unless necessary
Change desktop scrolling behavior for add image / add section image
tasks so that if the caption area is visible when pressing "yes",
there is no scrolling (previously the caption was scrolled near top
of the screen, and the image out of it, which was jarring). Do
scroll if the caption is off-screen (either because the image is
very tall, or because the user scrolled away, probably to examine
the article more closely).

Keep old behavior on mobile, where there is probably not enough
space to show both image and caption, and we also have to account
for the keyboard coming up, so scrolling the caption near top
makes sense.

Note that the scrolling behavior isn't super reliable, sometimes
when the image is completely off-screen, the screen jumps to it
immediately before animating. It doesn't always happen and I
couldn't figure out what triggers it (maybe the change in the
contenteditable? or maybe we have scrolling logic somewhere else),
but it's pre-existing behavior and I don't think the patch makes it
any worse.

Bug: T335209
Change-Id: Ide72e134526dd8c92c43b6c4721a80b37a22f4fa