Use CollationFactory in ConfigurationValidator for performance
I have seen this in a flame graph: Making a new collation is
apparently super expensive. The old code was doing this in advance the
moment the ConfigurationValidator service was instantiated via service
wiring, not knowing if it will ever be needed.

Instead inject the factory and make the collation only when we know we
need it. Later calls will be fast because the collations are cached in
the factory.

https: //performance.wikimedia.org/excimer/profile/ab497e55783487c4
Change-Id: Ia676dfe4bfdc9ac738972f1cd15b9bc17ad4c6c6