Section images: Select placeholder when inserting it
When inserting the suggested image placeholder into the VE
document, set the selection to an empty range right before the
placeholder (i.e. place the cursor there). The cursor location
is scrolled into view on some operations, so this was a source
of erratic scrolling behavior when clicking on the Yes button.

Bug: T335209
Change-Id: I7d959f84bb44d0b7b6c14f9418e58e7182ef4b6e