User impact: timezone cleanup
Why:
User timezone preference was initially used for
the computation of daily stats. This information
can be used to deanonymize users and it was removed
from the json stored in db and api responses in
T328643. Some classes are still user timezone aware
without a need.

What:
 - Remove the UserTimeCorrection object from the
edit data and user impact objects
 - Always use an empty user with the wiki default tz and
offset for computing the editCountByDay data

Bug: T329700
Change-Id: I9769dbc13ca30624340adb202721ceb5ccc8f44f