Mentor dashboard: Deprecate ApiSetMentorStatus
Why:

Originally, the mentor list was stored on a regular wiki page,
where it is not easily possible to store any other datapoints
than the mentor's username / message. For that reason, the away
timestamp was added to a preference, rather than the list of mentors,
where it would naturally belong to.

We introduced the structured list of mentors, where we can store
any information necessary for a mentor. While the actual mentor status
is not stored there yet (for no real reason; it should eventually change),
the structured mentor list brought a centralized API to manage the list
of mentors, action=growthmanagementorlist.

We should migrate away from action=growthsetmentorstatus and merge
it with the central API, which is what this patch does.

What:

* Add awayTimestampHuman to action=growthmanagementorlist,
  to make it easier to display the until timestamp in
  a human-readable format.
* Migrate to action=growthmanagementorlist in the Mentor tools
  Mentor dashboard module.
* Implement the "too long away" business case in MentorStatusManager
  rather than on the API level, as there is no reason why internal
  callers should set longer away timestamps than users can.
* Mark action=growthsetmentorstatus as deprecated.

Bug: T321503
Change-Id: Iad835a9c48e8fce18f4710680258683566f0a5b8