ChangeMentor: Refactor the notification conditions
The only case when notifications should be fired
is when the user has homepage/help panel enabled
AND their mentorship state is MENTORSHIP_ENABLED.
In all other cases, ChangeMentor should still
change the mentor assignment, but avoid firing
a notification for said event (as the users
probably aren't aware of mentorship).

Instead of having conditions that control
when notification is NOT send, formulate
a condition controlling when the notification
IS sent, as that's a smaller number of cases.

Also add better test coverage of when notification
should (not) be sent.

Bug: T336875
Change-Id: I1b34776e35058918853994d6884f29d2bc0855b9