Add reassignMentees.php maintenance script
Why:
Before the structured mentor list was deployed,
mentors removed themselves by only editing the wikitext
list of mentors. This caused many former mentors to have
mentees assigned (and ocasionally questions asked).

This was previously handled by a standalone Python
script, which was fine for one-off requests. Now,
we want to reassign mentees belonging to the ex-mentors
on all projects, and it is tedious to run the Python
script everywhere.

What:
Introduce `reassignMentees.php` maintenance script,
which is able to both find ex-mentors with assigned
mentees and reassign mentees belonging to them
to different mentors.

For pre-runs and individual cases in the future,
the possibility to reassign mentees assigned
to a given mentor is also provided.

This is somewhat based on https://phabricator.wikimedia.org/P37040.

Bug: T330071
Change-Id: Ie7545e14c48cee498d4bbc5f16cf1a324196ad80