IP Masking: Replace isRegistered/isAnon calls with isNamed()
Why:
The Growth team decided to put temporary users on par with anonymous
users. This means all checks that used to be "is this a registered
user" needs to now say "is this a named user/regular user".

In MW, there are two ways code could be checking for registereness:
via User::isRegistered and User::isAnon. To ensure temp. accounts are
on par with anonymous users, both checks need to be rewritten in terms
of User::isNamed.

What:
* Change calls to `User:isRegistered` to `User::isNamed`
* Change calls to `User::isAnon` to `!User::isNamed'.

For UserIdentity::isRegistered() and ::isAnon():

* Sometimes the object was actually a User and only parameter type
  declarations needed to be updated.
* Sometimes there was a User or Authority object nearby which could be
  used.
* If we have a UserFactory already injected, then this can be used to
  convert the UserIdentity to a User.
* Otherwise, inject the new UserIdentityUtils service. This was used in
  AddImageSubmissionHandler, AddLinkSubmissionHandler, ApiSetMentor,
  HomepageHooks, ImpactHooks, VisualEditorHooks.

Notes:
This patch should have no impact unless IP Masking is enabled. Snippet
for local testing:

```
$wgAutoCreateTempUser['enabled'] = true;
$wgGroupPermissions['temp']['edit'] = true;
```

It can be tested locally or on dewiki beta.

Bug: T341976
Depends-On: I94e8620cd3187f41f04cdf24b5b553baadfbc22d
Change-Id: I5e1517f64a1da043cdc7c9f3af6ed7b00a8c8fb6