Fix UserDatabaseHelper::hasMainspaceEdits
Prior to this patch, UserDatabaseHelper::hasMainspaceEdits
returned null if the user has more than 1000 of mainspace edits.
This is contrary to what it is supposed to do, ie. return null
if all of the first 1000 user edits are outside the main namespace.

Do this by removing the page_namespace condition (and fetch
first 1000 of user's edits, irrespective of their namespace)
and examine the values of page_namespace on the app level.

Regression test will come in a separate patch, to not block
the fix.

Follow-up: Ie3a46499342f6c494eb0d8da294809bab80922c8
Bug: T341994
Change-Id: Ia688668919d007d3a8dcca6b8e9a58cebecb2307