Do not acquire DB connections until needed
Lots of services are acquiring a DB connection in ServiceWiring.php.
This means that a working connection provider must already be present
by the time the ServiceWiring file is read, even if it's later never
used. This is generally the case in normal operations (i.e., serving web
responses), but not necessarily in other contexts, like database-less
maintenance scripts and PHPUnit tests.

For instance, core change I96ecf9ffd5fbf18db3e15e8774f7dc2fe5b9fd2a,
which disabled DB storage for database-less tests, is currently blocked
because many GE services are instantiated in hooks (and more generally
without the code under test doing anything explicit to instantiate
them). The services are effectively not used in those tests (or at
least, their DB access component is), but because this code tries to
acquire a connection upon instantiation, it fails immediately.

Also, constructors (and by extension wiring files) should be as
lightweight as possible; opening a DB connection is not lightweight.
Therefore, it would make sense to delay this until the connection is
actually used, even if the connection handles are lazy nowadays.

Inject IConnectionProvider or LoadBalancer instead, and obtain a
connection immediately before running queries.

Change-Id: I12083f165df8333f835be4cd5960d789ee6af409