api: Expose disabled task types in ApiInvalidateImageRecommendation
This API might be useful for external clients (such as the Wikimedia
Android app even when GrowthExperiments structured tasks are
disabled on a given wiki, and it's easy to always expose it.

Drop the check that the recommendation to be invalidatied really
belongs to the user. There isn't any real use case for it, and
it prevents external apps from selecting tasks for users in their
own way. AddImageFeedbackHandler already allows anyone to
invalidate an arbitrary recommendation so it doesn't make much
difference anyway,

Bug: T341150
Change-Id: I3c4c4c39eabfe2bc3c0a6d949a9cfbb5978acb42