PersonalizedPraiseSettings: Ignore non-int values
Why:
The settings form for Personalized praise module allows
for the numberic fields to be empty. If that happens, they're
sent to the server as an empty string. As of now, if empty state
is submitted, Mentor dashboard fully breaks (with a fatal error).

The setting comes from user properties, which have no guarantee
on the format. Even if the form validated an int is submitted,
it is a good idea to enforce this on the frontend (or rather,
not have any assumptions on the property value).

What:
Ensure PersonalizedPraiseSettings fetches an integer from the user
properties. If one is not available, cast it to an integer.

Bug: T343305
Change-Id: I14c827ee91b3d9c5635e889eb95141a952aefee3