Make WikiPageConfigLoader a no-op in non-database integration tests
For integration tests not in the Database group, we want to avoid DB
access via RevisionLookup.

WikiPageConfigLoader can be called in lots of other tests, because it is
accessed indirectly by things like getParamSettings() in API handlers
and hooks. Therefore, instead of requiring each potentially affected
test to mock it, just disable its lookup internally.

Ideally, this would use a different service class or something, but for
now, this solution should be sufficient.

Similarly, disable checks on the Title object in WikiPageConfig.

Bug: T155147
Change-Id: I3051bafa13a125da7e0639ff68cd9247e4a8ea8b