Structured mentor list: Use meaningful edit summaries
Why:
It is currently hard to understand what changes purely
by checking GrowthMentors.json's history, as the edit summaries
do not make it clear which action was taken by the user.

What:
Prefix the user-submitted reason by an informative prefix,
which makes it clear what the user actually did. Make use
of MW's autocomments, to avoid the need to pass around context
or build the comment on the UI level in all places where mentorship
related changes are needed. The autocomment is returned in content
language, as it might include user-provided strings (which presumably
are in content language).

Do the prefixing in StructuredMentorWriter (by calling
a newly-introduced MentorshipSummaryCreator class), so all
internal callers of mentorship features benefit from the summaries
without having to build them on their own.

Bug: T322115
Depends-On: I85f47d71f7f444f1c5b7bfea7897e4187d20d15c
Change-Id: I66008f7bfbf73b64c3c1c1a00131150dcef7852f