tests: Fix invalid user in HomepageModuleRegistryTest
This test either needs to use setService() to replace UserOptionsLookup
with a StaticUserOptionsLookup (and thus avoid the db failure from
DefaultOptionsLookup::verifyUsable due to the User object not being
backed by an actual user ID), in which case it can't cover the
ExperimentUserManager::setVariant() method and its call to setOption.

Or, it needs to go all in on being an integration test and use a real
User. It looks like the test is intending to cover integration with
the user preferences backend since it calls setVariant and setOption,
so I suggest the latter in this case.

This test was originally passing because verifyUsable()'s call to
isRegistered() was permitted to do a database lookup despite the test
lacking `@group Database`, because MediaWikiIntegrationTestCase does
not disable DB access.

This is changing soon in change I96ecf9ffd5fbf18 in MediaWiki core.
To prevent Growth from failing, core was changed to account for
unit tests without `@group Database in change I5987c4f1c3c502b23113,
which I am exploring to undo in change I8720d32fcfb18f7468e2 if I can
get extensions like this one to pass by either properly using the
DB or properly not using it, and not rely on e.g. a DB-related service
like DefaultOptionsLookup working in cases where DB is disabled,
which is something I don't think we should support.

Bug: T155147
Change-Id: I906450a4ac2dc51cdc536fbc978661b06f62f0a4